requirements.txt
requests
bs4
ebooklib
progress
pillow
simple_term_menu
